Created Lightbox + JS Fix
So, I decided to see if I could fix the issues with the Lightbox but after a while, I just couldn't get the code for the lightbox to work and I couldn't figure out why. So I gave up, I commented it out and got to work on my own.  Then, after I was done scripting, I had to test it. Nothing happen and I was left confused before realizing that the script tag for our javascript wasn't set up properly. So I fixed it and got the lightbox to work. The rest of my time was fixing the style and positions of the lightbox. I have left a little list in the CSS commented of my (or whoever wishes to work on it) to-do list.